# backend/docs/swagger/swagger.json
{
    "basePath": "/api/v1",
    "definitions": {
        "ActionItemResponse": {
            "properties": {
                "assessmentPeriod": {
                    "example": "2025 - 1st Half",
                    "type": "string"
                },
                "assignedTo": {
                    "example": "alice",
                    "type": "string"
                },
                "assigneeName": {
                    "example": "Alice Cooper",
                    "type": "string"
                },
                "createdAt": {
                    "example": "2026-09-08T10:21:33Z",
                    "type": "string"
                },
                "createdBy": {
                    "example": "teamlead1",
                    "type": "string"
                },
                "createdByName": {
                    "example": "Team Lead - Phoenix Squad",
                    "type": "string"
                },
                "description": {
                    "example": "Codebase health scored 2.6. Add tests to the payments and auth packages first.",
                    "type": "string"
                },
                "dimensionId": {
                    "example": "health",
                    "type": "string"
                },
                "dimensionName": {
                    "example": "Health of Codebase",
                    "type": "string"
                },
                "dueDate": {
                    "example": "2026-12-15T00:00:00Z",
                    "type": "string"
                },
                "id": {
                    "example": "c256e36d-a791-4474-a258-88aa25e6dc6f",
                    "type": "string"
                },
                "status": {
                    "example": "open",
                    "type": "string"
                },
                "teamId": {
                    "example": "team-phoenix",
                    "type": "string"
                },
                "title": {
                    "example": "Increase unit test coverage above 80%",
                    "type": "string"
                },
                "updatedAt": {
                    "example": "2026-09-08T10:21:33Z",
                    "type": "string"
                }
            },
            "type": "object"
        },
        "ActionItemsResponse": {
            "properties": {
                "actionItems": {
                    "items": {
                        "$ref": "#/definitions/ActionItemResponse"
                    },
                    "type": "array"
                }
            },
            "type": "object"
        },
...
        "CreateActionItemRequest": {
            "properties": {
                "assessmentPeriod": {
                    "example": "2025 - 1st Half",
                    "type": "string"
                },
                "assignedTo": {
                    "example": "alice",
                    "type": "string"
                },
                "description": {
                    "example": "Codebase health scored 2.6. Add tests to the payments and auth packages first.",
                    "type": "string"
                },
                "dimensionId": {
                    "example": "health",
                    "type": "string"
                },
                "dueDate": {
                    "description": "ISO date string YYYY-MM-DD",
                    "example": "2026-12-15",
                    "type": "string"
                },
                "title": {
                    "example": "Increase unit test coverage above 80%",
                    "maxLength": 500,
                    "type": "string"
                }
            },
            "required": [
                "title"
            ],
            "type": "object"
        },
...
        "ErrorResponse": {
            "properties": {
                "code": {
                    "example": "TEAM_NOT_FOUND",
                    "type": "string"
                },
                "error": {
                    "example": "Team not found",
                    "type": "string"
                },
                "message": {
                    "example": "No team exists with id team-unknown",
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "TeamActionSummaryResponse": {
            "properties": {
                "openCount": {
                    "example": 1,
                    "type": "integer"
                },
                "teamId": {
                    "example": "team-phoenix",
                    "type": "string"
                },
                "teamName": {
                    "example": "Phoenix Squad",
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "TeamsActionSummaryResponse": {
            "properties": {
                "teams": {
                    "items": {
                        "$ref": "#/definitions/TeamActionSummaryResponse"
                    },
                    "type": "array"
                }
            },
            "type": "object"
        },
...
        "UpdateActionItemRequest": {
            "properties": {
                "assessmentPeriod": {
                    "example": "2025 - 1st Half",
                    "type": "string"
                },
                "assignedTo": {
                    "example": "alice",
                    "type": "string"
                },
                "description": {
                    "example": "Codebase health scored 2.6. Add tests to the payments and auth packages first.",
                    "type": "string"
                },
                "dimensionId": {
                    "example": "health",
                    "type": "string"
                },
                "dueDate": {
                    "example": "2026-12-15",
                    "type": "string"
                },
                "status": {
                    "example": "in_progress",
                    "type": "string"
                },
                "title": {
                    "example": "Increase unit test coverage above 80%",
                    "maxLength": 500,
                    "type": "string"
                }
            },
            "type": "object"
        },
...
    },
    "info": {
        "contact": {},
        "description": "Team360 Health Check Application REST API. Assesses team health across 11 dimensions with hierarchical org support.",
        "title": "Team360 API",
        "version": "1.0"
    },
    "paths": {
...
        "/managers/{managerId}/teams/action-items": {
            "get": {
                "description": "Returns, for every team the given manager supervises, the team's ID, name, and count of action items whose status is not \"done\", ordered by team name. Access requires a manager-or-above role, and the authenticated caller may only request their own summary; requesting another manager's summary is rejected with a 403.",
                "parameters": [
                    {
                        "description": "Manager ID",
                        "in": "path",
                        "name": "managerId",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TeamsActionSummaryResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — can only view your own summary",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch/read action summaries",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get action item counts by team",
                "tags": [
                    "Action Items"
                ]
            }
        },
...
        "/teams/{teamId}/action-items": {
            "get": {
                "description": "Returns the action items belonging to the given team, newest first, optionally narrowed by status (open, in_progress, done) and/or assessment period via query parameters. Each item includes its dimension, creator, assignee, title, description, status, due date, assessment period, and timestamps, with names resolved for the linked dimension, creator, and assignee. Access requires membership on the team.",
                "parameters": [
                    {
                        "description": "Team ID",
                        "in": "path",
                        "name": "teamId",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Filter by status (open, in_progress, done)",
                        "in": "query",
                        "name": "status",
                        "type": "string"
                    },
                    {
                        "description": "Filter by assessment period",
                        "in": "query",
                        "name": "period",
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ActionItemsResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch/scan/read action items",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List a team's action items",
                "tags": [
                    "Action Items"
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "description": "Creates a new action item for the given team with a title, description, and optional dimension, assignee, due date, and assessment period. New items are always created with status \"open\". When assignedTo is set, the target user must already be a member of the team or the request is rejected with a 400. Access requires membership on the team.",
                "parameters": [
                    {
                        "description": "Team ID",
                        "in": "path",
                        "name": "teamId",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Action item",
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateActionItemRequest"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "id, status, createdAt",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid request, invalid dueDate, or assignedTo not a team member",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to validate assignee or create action item",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Create a team action item",
                "tags": [
                    "Action Items"
                ]
            }
        },
        "/teams/{teamId}/action-items/{id}": {
            "delete": {
                "description": "Permanently deletes the action item identified by the path ID from the given team. Returns a 404 if no matching item exists for that team. Access requires membership on the team.",
                "parameters": [
                    {
                        "description": "Team ID",
                        "in": "path",
                        "name": "teamId",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Action item ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "deleted",
                        "schema": {
                            "additionalProperties": {
                                "type": "boolean"
                            },
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Action item not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete action item",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Delete a team action item",
                "tags": [
                    "Action Items"
                ]
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "description": "Partially updates the action item identified by the path ID within the given team; only the fields present in the request body are changed. Status, when provided, must be one of open, in_progress, or done. When assignedTo is set, the target user must already be a member of the team or the request is rejected with a 400. Returns a 404 if no matching item exists for that team. Access requires membership on the team.",
                "parameters": [
                    {
                        "description": "Team ID",
                        "in": "path",
                        "name": "teamId",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Action item ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Fields to update",
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateActionItemRequest"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "updated",
                        "schema": {
                            "additionalProperties": {
                                "type": "boolean"
                            },
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid request, invalid status/dueDate, or assignedTo not a team member",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Action item not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to validate assignee or update action item",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Update a team action item",
                "tags": [
                    "Action Items"
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
    "swagger": "2.0"
}